Compile WASM plugins at test time, not during regular builds
