+++ Cargo.toml
@@ -58,6 +58,7 @@ xdg = "2.5"
 
 [dev-dependencies]
 assert_cmd = "2.0"
+ctor = "1.0"
 predicates = "2.0"
 similar = "2.4"
 shlex = "1.3.0"
