+++ build.rs
@@ -16,7 +16,6 @@ fn main() {
     fs::create_dir_all(&generated_tests_dir).unwrap();
 
     compile_test_grammars(Path::new(&out_dir));
-    build_wasm_plugins(Path::new(&out_dir));
 
     // Empty plugins dir for tests to use to isolate themselves from the host
     // system default plugins dir
@@ -47,190 +46,6 @@ fn main() {
     generate_tests_include_file(&generated_tests_dir, &test_cases);
 
     println!("cargo:rerun-if-changed=tests/test_cases");
-    println!("cargo:rerun-if-changed=plugins/common");
-}
-
-/// Build plugins and make it possible to use them in integration tests
-///
-/// Discovers every subdirectory of `plugins/` that has both a `Cargo.toml` and a
-/// `plugin.toml` template, compiles each to a `.wasm` component targeting
-/// `wasm32-wasip2`, and writes the result to `$OUT_DIR/plugins/{plugin_name}/`.
-/// A single env var `TREETAGS_TEST_PLUGINS_DIR` is emitted pointing at that
-/// directory so integration tests can pass `--plugin-dir {TREETAGS_TEST_PLUGINS_DIR}`
-/// without embedding any absolute paths in the checked-in source tree.
-fn build_wasm_plugins(out_dir: &Path) {
-    let target_available = std::process::Command::new("rustup")
-        .args(["target", "list", "--installed"])
-        .output()
-        .map(|o| String::from_utf8_lossy(&o.stdout).contains("wasm32-wasip2"))
-        .unwrap_or(false);
-
-    if !target_available {
-        println!("cargo:warning=wasm32-wasip2 target not installed; skipping plugin compilation. Run: rustup target add wasm32-wasip2");
-        return;
-    }
-
-    let plugins_src_dir = Path::new("plugins");
-    if !plugins_src_dir.exists() {
-        return;
-    }
-
-    let plugins_out_dir = out_dir.join("plugins");
-    // Use a separate target dir to avoid lock contention with the parent cargo invocation.
-    let wasm_target_dir = out_dir.join("wasm-target");
-    let cargo = env::var("CARGO").unwrap_or_else(|_| "cargo".to_string());
-    let mut any_built = false;
-
-    let mut entries: Vec<_> = fs::read_dir(plugins_src_dir)
-        .expect("cannot read plugins directory")
-        .filter_map(Result::ok)
-        .collect();
-    entries.sort_by_key(|e| e.path());
-
-    for entry in entries {
-        let plugin_src = entry.path();
-        if !plugin_src.is_dir() {
-            continue;
-        }
-
-        let cargo_toml_path = plugin_src.join("Cargo.toml");
-        let manifest_template_path = plugin_src.join("plugin.toml");
-        if !cargo_toml_path.exists() || !manifest_template_path.exists() {
-            continue;
-        }
-
-        let cargo_toml = match fs::read_to_string(&cargo_toml_path) {
-            Ok(s) => s,
-            Err(e) => {
-                println!("cargo:warning=cannot read {cargo_toml_path:?}: {e}");
-                continue;
-            }
-        };
-        let crate_name = match parse_cargo_package_name(&cargo_toml) {
-            Some(n) => n,
-            None => {
-                println!("cargo:warning=no package name found in {cargo_toml_path:?}");
-                continue;
-            }
-        };
-
-        let manifest_template = match fs::read_to_string(&manifest_template_path) {
-            Ok(s) => s,
-            Err(e) => {
-                println!("cargo:warning=cannot read {manifest_template_path:?}: {e}");
-                continue;
-            }
-        };
-        let plugin_name = match parse_toml_string_field(&manifest_template, "name") {
-            Some(n) => n,
-            None => {
-                println!("cargo:warning=no 'name' field in {manifest_template_path:?}");
-                continue;
-            }
-        };
-
-        let mut cmd = std::process::Command::new(&cargo);
-        cmd.args([
-            "build",
-            "--target",
-            "wasm32-wasip2",
-            "--release",
-            "-p",
-            &crate_name,
-        ])
-        .env("CARGO_TARGET_DIR", &wasm_target_dir);
-
-        let status = cmd.status();
-
-        let wasm_filename = format!("{}.wasm", crate_name.replace('-', "_"));
-        let wasm_src = wasm_target_dir
-            .join("wasm32-wasip2/release")
-            .join(&wasm_filename);
-
-        match status {
-            Ok(s) if s.success() && wasm_src.exists() => {}
-            Ok(s) => {
-                println!(
-                    "cargo:warning={plugin_name} plugin compilation failed (exit {:?}); tests will be skipped",
-                    s.code()
-                );
-                continue;
-            }
-            Err(e) => {
-                println!(
-                    "cargo:warning=could not invoke cargo for {plugin_name} plugin: {e}; tests will be skipped"
-                );
-                continue;
-            }
-        }
-
-        let plugin_out_dir = plugins_out_dir.join(&plugin_name);
-        if let Err(e) = fs::create_dir_all(&plugin_out_dir) {
-            println!("cargo:warning=cannot create {plugin_out_dir:?}: {e}");
-            continue;
-        }
-
-        let wasm_out_path = plugin_out_dir.join("plugin.wasm");
-        if let Err(e) = fs::copy(&wasm_src, &wasm_out_path) {
-            println!("cargo:warning=cannot copy {plugin_name} .wasm: {e}");
-            continue;
-        }
-
-        // Prepend `wasm_file` as a top-level key. Appending it would fold the
-        // key into a trailing table section (e.g. `[[kinds]]` or
-        // `[disambiguation]`) and corrupt the manifest.
-        let manifest = format!(
-            "wasm_file = \"{}\"\n{}\n",
-            wasm_out_path.display(),
-            manifest_template.trim_start()
-        );
-        if let Err(e) = fs::write(plugin_out_dir.join("plugin.toml"), &manifest) {
-            println!("cargo:warning=cannot write {plugin_name} plugin.toml: {e}");
-            continue;
-        }
-
-        println!("cargo:rerun-if-changed={}", plugin_src.display());
-        any_built = true;
-    }
-
-    if any_built {
-        println!(
-            "cargo:rustc-env=TREETAGS_TEST_PLUGINS_DIR={}",
-            plugins_out_dir.display()
-        );
-    }
-}
-
-fn parse_cargo_package_name(content: &str) -> Option<String> {
-    let mut in_package = false;
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed == "[package]" {
-            in_package = true;
-            continue;
-        }
-        if in_package && trimmed.starts_with('[') {
-            break;
-        }
-        if in_package && trimmed.starts_with("name") {
-            if let Some(val) = trimmed.splitn(2, '=').nth(1) {
-                return Some(val.trim().trim_matches('"').to_string());
-            }
-        }
-    }
-    None
-}
-
-fn parse_toml_string_field(content: &str, field: &str) -> Option<String> {
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with(field) {
-            if let Some(val) = trimmed.splitn(2, '=').nth(1) {
-                return Some(val.trim().trim_matches('"').to_string());
-            }
-        }
-    }
-    None
 }
 
 #[derive(Debug, Clone)]
