+++ tests/helpers/mod.rs
@@ -1,3 +1,4 @@
 pub mod file_utils;
 pub mod golden_test_runner;
+pub mod plugin_builder;
 pub mod test_runner;
